move fixtures to conftest and remove extraneous imports
